solve the form reload.

+++ Task-9/app.js
@@ -1,18 +1,25 @@
-$('.datepicker').pickadate({
-    disable: [1]
+$(document).ready(function() {
+    $('.datepicker').pickadate({
+        disable: [7],
+        required: true,
+        firstDay: "Monday",
+        editable: true,
+        min: true
+    })
 })
 
-let noData = `<tr class="table-dark"><td colspan='9' class='text-center'>No Records Available</td></tr>`;
+
+var noData = "<tr class='table-dark'><td colspan='9' class='text-center'> No Records Available</td></tr>";
 
 function addEmptyRow() {
-    
+
     if ($("#tblData tbody").children().children().length == 0) {
         $("#tblData tbody").append(noData);
     }
 }
 
 
-function clearForm() {
+function clearForm() { 
     $("#txtName").val("");
     $("#assignedTask").val("");
     $("#initDate").val("");
@@ -23,10 +30,9 @@ function clearForm() {
 
 
 
-function buildTable(data) {
+function buildTable(data) {  
     $("#tblData tbody").html("");
-    let i = 1;
-    let bgColor = "";
+    var i = 1;
     data.forEach(element => {
         if (element.status == "In-progress") {
             bgColor = "table-primary";
@@ -39,14 +45,14 @@ function buildTable(data) {
         let tableRow = `
           <tr class='${bgColor}'>
           <td class='text-center'> ${i}</td>
-          <td class='txtName' data-id="${element.id}">${element.personName} </td>
+          <td class='txtName'  data-id='${element.id}'>${element.personName} </td>
           <td class='assignedTask'>${element.assignedTask}</td>
           <td class='initDate'>${element.initDate}</td>
           <td class='dueDate'>${element.dueDate}</td>
           <td class='comments'>${element.comments}</td>
           <td class='status'>${element.status}</td>
           <td class='text-center'>
-          <button type="button" class='btn btn-sm btn-success btn-edit' data-bs-toggle="modal" data-bs-target="#staticBackdrop">Edit</button>
+          <button type='button' class='btn btn-sm btn-success btn-edit' data-bs-toggle='modal' data-bs-target='#staticBackdrop'>Edit</button>
           <button class='btn btn-sm btn-danger btn-delete'>Delete</button>
           </td>
         </tr>`
@@ -55,7 +61,7 @@ function buildTable(data) {
     });
 }
 
-function loadDataFromLocal() {
+function loadDataFromLocal() {  
     let localData = localStorage.getItem('localData');
     let localArray = JSON.parse(localData);
     if (localData) {
@@ -64,7 +70,7 @@ function loadDataFromLocal() {
     addEmptyRow();
 }
 
-function addDataToLocal() {
+function addDataToLocal() {   
     let localData = localStorage.getItem('localData');
 
     if (localData) {
@@ -82,11 +88,28 @@ function addDataToLocal() {
         localArray.push(obj);
         localStorage.setItem('localData', JSON.stringify(localArray));
         loadDataFromLocal();
+    } else {
+        const arryObj = [];
+        const obj = {
+            id: 1,
+            personName: $("#txtName").val(),
+            assignedTask: $("#assignedTask").val(),
+            initDate: $("#initDate").val(),
+            dueDate: $("#dueDate").val(),
+            comments: $("#comments").val(),
+            status: $("#status").val()
+        };
+        arryObj.push(obj);
+        localStorage.setItem('localData', JSON.stringify(arryObj));
+        loadDataFromLocal();
     }
     clearForm();
 }
 
-function updateDataFromLocal() {
+
+
+
+function updateDataFromLocal() {  
     let localData = localStorage.getItem('localData');
     let localArray = JSON.parse(localData);
     const oldRecord = localArray.find(m => m.id == $("#txtId").val());
@@ -99,9 +122,10 @@ function updateDataFromLocal() {
     localStorage.setItem('localData', JSON.stringify(localArray));
     loadDataFromLocal();
     clearForm();
+
 }
 
-function deleteDataFromLocal(id) {
+function deleteDataFromLocal(id) { 
     let localData = localStorage.getItem('localData');
     let localArray = JSON.parse(localData);
     let i = 0;
@@ -115,7 +139,7 @@ function deleteDataFromLocal(id) {
     localStorage.setItem('localData', JSON.stringify(localArray));
     loadDataFromLocal();
 }
-$(document).ready(function() {
+$(document).ready(function() { 
     loadDataFromLocal();
     $('#tblData').on('click', '.btn-edit', function() {
         const personName = $(this).parent().parent().find(".txtName").html();
@@ -133,27 +157,51 @@ $(document).ready(function() {
         $("#txtId").val(id);
         $("#status").val(status);
         $("#btnSave").text("Update");
+
     });
+});
 
-    $('#tblData').on('click', '.btn-delete', function() {
-        
-        const id = $(this).parent().parent().find(".txtName").attr("data-id");
-        deleteDataFromLocal(id);
-    });
-    $('.newEntry').on('click', function() {
-        clearForm();
-    })
-
-    $("#form").on('submit', function() {
-        if ($("#txtId").val() == '') {
-            addDataToLocal();
-        } else {
-
-            updateDataFromLocal();
+$('#tblData').on('click', '.btn-delete', function() { 
+    swal({
+        title: "Are you sure?",
+        text: "you want to delete!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+    }).then((willDelete) => {
+        if (willDelete) {
+            const id = $(this).parent().parent().find(".txtName").attr("data-id");
+            deleteDataFromLocal(id);
+            swal("Your information has been deleted successfully", {
+                icon: "success",
+            });
         }
-          $('#staticBackdrop').modal('hide');
     });
-    $("#btnClear").click(function() {
-        clearForm();
-    });
+});
+
+
+$('.newEntry').on('click', function() { 
+    clearForm();
+});
+$("#form").on('submit', function(e) { 
+    if ($("#txtId").val() == '') {
+        addDataToLocal();
+
+        var dataString = $(this).serialize();
+        $.ajax({
+            type: "POST",
+            url: "index.html"
+        });
+
+        e.preventDefault();
+    } else {
+
+        updateDataFromLocal();
+
+    }
+
+    $('#staticBackdrop').modal('hide');
+});
+$("#btnClear").click(function() {
+    clearForm();
 });
